# contracts/openapi/teamos.yaml
openapi: 3.1.0
...
info:
  title: TeamOS Gateway API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Внешний JSON/REST-контракт TeamOS. Поля и формы ответов зеркалят
    `src/api/index.ts` и `src/types/index.ts` frontend-приложения.
    Все даты-время передаются в ISO-8601, идентификаторы являются UUID.
servers:
  - url: http://localhost:8080
    description: Локальный gateway
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/kb/sections:
    get:
      tags: [Knowledge Base]
      operationId: getSections
      summary: Получить разделы базы знаний
      responses:
        '200':
          $ref: '#/components/responses/ArticleSections'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Knowledge Base]
      operationId: createSection
      summary: Создать раздел базы знаний
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleSectionInput'
      responses:
        '201':
          $ref: '#/components/responses/ArticleSection'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/kb/sections/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Knowledge Base]
      operationId: updateSection
      summary: Обновить раздел базы знаний
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleSectionInput'
      responses:
        '200':
          $ref: '#/components/responses/ArticleSection'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Knowledge Base]
      operationId: deleteSection
      summary: Удалить пустой раздел базы знаний
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/kb/articles:
    get:
      tags: [Knowledge Base]
      operationId: getArticles
      summary: Получить статьи
      parameters:
        - name: sectionId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          $ref: '#/components/responses/Articles'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Knowledge Base]
      operationId: createArticle
      summary: Создать статью
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleInput'
      responses:
        '201':
          $ref: '#/components/responses/Article'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/kb/articles/search:
    get:
      tags: [Knowledge Base]
      operationId: searchArticles
      summary: Найти статьи по заголовку и содержимому
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          $ref: '#/components/responses/Articles'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/kb/articles/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Knowledge Base]
      operationId: getArticle
      summary: Получить статью
      responses:
        '200':
          $ref: '#/components/responses/Article'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Knowledge Base]
      operationId: updateArticle
      summary: Обновить статью
      parameters:
        - name: If-Match
          in: header
          required: false
          description: Ожидаемый номер версии для optimistic locking.
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleInput'
      responses:
        '200':
          $ref: '#/components/responses/Article'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/kb/articles/{articleId}/rollback:
    parameters:
      - name: articleId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ID'
    post:
      tags: [Knowledge Base]
      operationId: rollbackArticle
      summary: Откатить статью к сохранённой версии
      parameters:
        - name: If-Match
          in: header
          required: false
          description: Ожидаемый номер текущей версии для optimistic locking.
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackArticleInput'
      responses:
        '200':
          $ref: '#/components/responses/Article'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/kb/articles/{id}/versions:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Knowledge Base]
      operationId: getArticleVersions
      summary: Получить историю версий статьи
      responses:
        '200':
          $ref: '#/components/responses/ArticleVersions'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/kb/articles/{id}/acknowledgements:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Knowledge Base]
      operationId: getAcknowledgements
      summary: Получить отметки об ознакомлении со статьёй
      responses:
        '200':
          $ref: '#/components/responses/Acknowledgements'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/kb/articles/{id}/acknowledge:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Knowledge Base]
      operationId: acknowledgeArticle
      summary: Отметить статью как прочитанную
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      operationId: getNotifications
      summary: Получить уведомления текущего пользователя
      responses:
        '200':
          $ref: '#/components/responses/Notifications'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/notifications/unread-count:
    get:
      tags: [Notifications]
      operationId: getUnreadCount
      summary: Получить число непрочитанных уведомлений
      responses:
        '200':
          description: Число непрочитанных уведомлений
          content:
            application/json:
              schema:
                type: integer
                minimum: 0
        default:
          $ref: '#/components/responses/Error'
  /api/v1/notifications/{id}/read:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Notifications]
      operationId: markRead
      summary: Отметить уведомление прочитанным
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/notifications/read-all:
    post:
      tags: [Notifications]
      operationId: markAllRead
      summary: Отметить все уведомления прочитанными
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/notifications/stream:
    get:
      tags: [Notifications]
      operationId: streamNotifications
      summary: SSE-поток новых уведомлений
      responses:
        '200':
          description: Поток Server-Sent Events с event type `notification`.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-cache
          content:
            text/event-stream:
              schema:
                type: string
                examples:
                  - |-
                    event: notification
                    data: {"id":"4dc6f749-3490-4fba-a65b-753f747c5473"}
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: teamos_refresh
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ID'
...
  responses:
    NoContent:
      description: Операция выполнена
...
    Conflict:
      description: Конфликт уникальности, состояния или версии
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error:
      description: Ошибка API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ArticleSection:
      description: Раздел базы знаний
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArticleSection'
    ArticleSections:
      description: Разделы базы знаний
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ArticleSection'
    Article:
      description: Статья
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Article'
    Articles:
      description: Статьи
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Article'
    ArticleVersions:
      description: Версии статьи
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ArticleVersion'
    Acknowledgements:
      description: Отметки об ознакомлении
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Acknowledgement'
...
    Notifications:
      description: Уведомления
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/AppNotification'
...
  schemas:
    ID:
      type: string
      format: uuid
    ISODateTime:
      type: string
      format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [message, status]
          properties:
            message:
              type: string
              minLength: 1
              example: Сотрудник не найден
            status:
              type: integer
              minimum: 400
              maximum: 599
              example: 404
    RichTextContent:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          type: string
          const: doc
        content:
          type: array
          items: {}
...
    AccessScope:
      type: string
      enum: [company, custom]
    AccessSettings:
      type: object
      additionalProperties: false
      required: [scope, departmentIds, positionIds, userIds]
      properties:
        scope:
          $ref: '#/components/schemas/AccessScope'
        departmentIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        positionIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        userIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
    ArticleSection:
      type: object
      additionalProperties: false
      required: [id, name, parentId, order, access]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
          minLength: 1
        parentId:
          type: [string, 'null']
          format: uuid
          x-go-type: nullable.Nullable[string]
          x-go-type-skip-optional-pointer: true
        order:
          type: integer
          minimum: 0
        access:
          $ref: '#/components/schemas/AccessSettings'
    ArticleStatus:
      type: string
      enum: [draft, published]
    Article:
      type: object
      additionalProperties: false
      required:
        - id
        - sectionId
        - title
        - content
        - status
        - authorId
        - version
        - requiresAcknowledgement
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/ID'
        sectionId:
          $ref: '#/components/schemas/ID'
        title:
          type: string
          minLength: 1
        content:
          $ref: '#/components/schemas/RichTextContent'
        status:
          $ref: '#/components/schemas/ArticleStatus'
        authorId:
          $ref: '#/components/schemas/ID'
        version:
          type: integer
          minimum: 1
        requiresAcknowledgement:
          type: boolean
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
        updatedAt:
          $ref: '#/components/schemas/ISODateTime'
    ArticleVersion:
      type: object
      additionalProperties: false
      required: [id, articleId, version, title, content, authorId, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        articleId:
          $ref: '#/components/schemas/ID'
        version:
          type: integer
          minimum: 1
        title:
          type: string
        content:
          $ref: '#/components/schemas/RichTextContent'
        authorId:
          $ref: '#/components/schemas/ID'
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
    Acknowledgement:
      type: object
      additionalProperties: false
      required: [articleId, userId, acknowledgedAt]
      properties:
        articleId:
          $ref: '#/components/schemas/ID'
        userId:
          $ref: '#/components/schemas/ID'
        acknowledgedAt:
          $ref: '#/components/schemas/ISODateTime'
    CreateArticleSectionInput:
      type: object
      additionalProperties: false
      required: [name, parentId]
      properties:
        name:
          type: string
          minLength: 1
        parentId:
          type: [string, 'null']
          format: uuid
          x-go-type: nullable.Nullable[string]
          x-go-type-skip-optional-pointer: true
        access:
          $ref: '#/components/schemas/AccessSettings'
    UpdateArticleSectionInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        access:
          $ref: '#/components/schemas/AccessSettings'
    CreateArticleInput:
      type: object
      additionalProperties: false
      required: [sectionId, title, content, status, requiresAcknowledgement]
      properties:
        sectionId:
          $ref: '#/components/schemas/ID'
        title:
          type: string
          minLength: 1
        content:
          $ref: '#/components/schemas/RichTextContent'
        status:
          $ref: '#/components/schemas/ArticleStatus'
        requiresAcknowledgement:
          type: boolean
    UpdateArticleInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        sectionId:
          $ref: '#/components/schemas/ID'
        title:
          type: string
          minLength: 1
        content:
          $ref: '#/components/schemas/RichTextContent'
        status:
          $ref: '#/components/schemas/ArticleStatus'
        requiresAcknowledgement:
          type: boolean
    RollbackArticleInput:
      type: object
      additionalProperties: false
      required: [versionId]
      properties:
        versionId:
          $ref: '#/components/schemas/ID'
...
    NotificationType:
      type: string
      enum:
        - task_assigned
        - task_comment
        - task_due
        - article_published
        - article_ack_required
        - course_assigned
        - course_due
        - mention
    AppNotification:
      type: object
      additionalProperties: false
      required: [id, userId, type, title, read, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        userId:
          $ref: '#/components/schemas/ID'
        type:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
        body:
          type: string
        link:
          type: string
        read:
          type: boolean
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
...