# services/kb/internal/storage/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Acknowledgement struct {
	CompanyID      uuid.UUID `json:"company_id"`
	ArticleID      uuid.UUID `json:"article_id"`
	UserID         uuid.UUID `json:"user_id"`
	AcknowledgedAt time.Time `json:"acknowledged_at"`
}

type Article struct {
	ID                      uuid.UUID   `json:"id"`
	CompanyID               uuid.UUID   `json:"company_id"`
	SectionID               uuid.UUID   `json:"section_id"`
	Title                   string      `json:"title"`
	Content                 []byte      `json:"content"`
	Status                  string      `json:"status"`
	AuthorID                uuid.UUID   `json:"author_id"`
	Version                 int32       `json:"version"`
	RequiresAcknowledgement bool        `json:"requires_acknowledgement"`
	PlainText               string      `json:"plain_text"`
	Search                  interface{} `json:"search"`
	CreatedAt               time.Time   `json:"created_at"`
	UpdatedAt               time.Time   `json:"updated_at"`
}

type ArticleVersion struct {
	ID        uuid.UUID `json:"id"`
	CompanyID uuid.UUID `json:"company_id"`
	ArticleID uuid.UUID `json:"article_id"`
	Version   int32     `json:"version"`
	Title     string    `json:"title"`
	Content   []byte    `json:"content"`
	AuthorID  uuid.UUID `json:"author_id"`
	CreatedAt time.Time `json:"created_at"`
}

type Outbox struct {
	ID            uuid.UUID          `json:"id"`
	CompanyID     uuid.UUID          `json:"company_id"`
	AggregateID   uuid.UUID          `json:"aggregate_id"`
	EventOrder    int64              `json:"event_order"`
	Subject       string             `json:"subject"`
	Payload       []byte             `json:"payload"`
	Headers       []byte             `json:"headers"`
	OccurredAt    time.Time          `json:"occurred_at"`
	NextAttemptAt time.Time          `json:"next_attempt_at"`
	PublishedAt   pgtype.Timestamptz `json:"published_at"`
	Attempts      int32              `json:"attempts"`
	LastError     pgtype.Text        `json:"last_error"`
}

type ProcessedEvent struct {
	EventID     uuid.UUID `json:"event_id"`
	CompanyID   uuid.UUID `json:"company_id"`
	ProcessedAt time.Time `json:"processed_at"`
}

type Section struct {
	ID        uuid.UUID     `json:"id"`
	CompanyID uuid.UUID     `json:"company_id"`
	Name      string        `json:"name"`
	ParentID  uuid.NullUUID `json:"parent_id"`
	Order     int32         `json:"order"`
	Access    []byte        `json:"access"`
	CreatedAt time.Time     `json:"created_at"`
	UpdatedAt time.Time     `json:"updated_at"`
}